requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
hazelcast-python-client==5.3.0